requirements.txt
ragas
langchain
langchain-community
langchain_openai
python-dotenv
chromadb
cohere
langchainhub
streamlit
tqdm
"unstructured[all-docs]"
st-pages
plotly
"git+https://github.com/misbahsy/rerankers.git#egg=rerankers[all]"
langchain-anthropic
langchain-google-genai